Rename in_app_purchase directories and update imports
