+++ lib/main.dart
@@ -14,7 +14,7 @@ import 'package:naijasingles/common/routes/router.dart';
 import 'package:naijasingles/features/auth/facebook_login/facebook_login_bloc.dart';
 import 'package:naijasingles/features/explore/bloc/explore_map_bloc.dart';
 import 'package:naijasingles/features/home/ui/screens/splash.dart';
-import 'package:naijasingles/features/payment/ui/in-app%20purcahse/get%20Products/getproducts_bloc.dart';
+import 'package:naijasingles/features/payment/ui/in_app_purchase/get_products/getproducts_bloc.dart';
 import 'package:naijasingles/features/street_view/bloc/streetviewdata_bloc.dart';
 import 'package:naijasingles/services/location/bloc/userlocation_bloc.dart';
 import 'package:in_app_purchase_android/in_app_purchase_android.dart';
@@ -32,7 +32,7 @@ import 'features/home/bloc/searchuser_bloc.dart';
 import 'features/home/bloc/swipebloc_bloc.dart';
 import 'features/home/ui/screens/user_filter/bloc/userfilter_bloc.dart';
 import 'features/match/bloc/match_bloc.dart';
-import 'features/payment/ui/in-app purcahse/buy Products/buyproducts_bloc.dart';
+import 'features/payment/ui/in_app_purchase/buy_products/buyproducts_bloc.dart';
 import 'features/report/bloc/report_bloc.dart';
 import 'features/user/bloc/update_user_bloc.dart';
 import 'firebase_options.dart';
